+++ Controls-demo/list_new/Grouped/CollapsedGroups/Index.ts
@@ -1,6 +1,6 @@
 import {Control, TemplateFunction, IControlOptions} from 'UI/Base';
 import {Memory} from 'Types/source';
-import {getGroupedCatalog as getData} from '../../DemoHelpers/DataCatalog';
+import {getGroupedCatalog as getData} from '../../DemoHelpers/Data/Groups';
 import * as Template from 'wml!Controls-demo/list_new/Grouped/CollapsedGroups/CollapsedGroups';
 
 export default class extends Control<IControlOptions> {
