+++ Controls-demo/list_new/Grouped/HiddenGroup/Index.ts
@@ -1,6 +1,6 @@
 import {Control, TemplateFunction, IControlOptions} from 'UI/Base';
 import {Memory} from 'Types/source';
-import {getGroupedCatalogWithHiddenGroup as getData} from '../../data/HiddenGroup';
+import {getGroupedCatalogWithHiddenGroup as getData} from '../../DemoHelpers/Data/HiddenGroup';
 import * as Template from 'wml!Controls-demo/list_new/Grouped/HiddenGroup/HiddenGroup';
 
 export default class extends Control<IControlOptions> {
