+++ Controls-demo/list_new/Grouped/HiddenGroupSwitchingGroup/Index.ts
@@ -1,7 +1,7 @@
 import {Control, TemplateFunction, IControlOptions} from 'UI/Base';
 import {Memory} from 'Types/source';
 import {groupConstants} from 'Controls/list';
-import {getGroupedCatalogForSwitchingGroup as getData} from '../../data/HiddenGroup';
+import {getGroupedCatalogForSwitchingGroup as getData} from '../../DemoHelpers/Data/HiddenGroup';
 import * as Template from 'wml!Controls-demo/list_new/Grouped/HiddenGroupSwitchingGroup/HiddenGroupSwitchingGroup';
 
 export default class extends Control<IControlOptions> {
